Device Plugin fake-node mode support
