+++ internal/deviceplugin/device_plugin.go
@@ -1,229 +1,34 @@
 package deviceplugin
 
 import (
-	"fmt"
-	"log"
-	"net"
-	"os"
-	"path"
-	"strings"
-	"time"
-
-	"github.com/google/uuid"
+	"github.com/run-ai/fake-gpu-operator/internal/common/constants"
 	"github.com/run-ai/fake-gpu-operator/internal/common/topology"
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-	pluginapi "k8s.io/kubelet/pkg/apis/deviceplugin/v1beta1"
+	"github.com/spf13/viper"
+	"k8s.io/client-go/kubernetes"
 )
 
 const (
 	resourceName = "nvidia.com/gpu"
-	serverSock   = pluginapi.DevicePluginPath + "fake-nvidia-gpu.sock"
 )
 
-type DevicePlugin struct {
-	devs   []*pluginapi.Device
-	socket string
-
-	stop   chan interface{}
-	health chan *pluginapi.Device
-	server *grpc.Server
+type Interface interface {
+	Serve() error
 }
 
-func NewDevicePlugin(topology *topology.NodeTopology) *DevicePlugin {
+func NewDevicePlugin(topology *topology.NodeTopology, kubeClient kubernetes.Interface) Interface {
 	if topology == nil {
 		panic("topology is nil")
 	}
 
-	return &DevicePlugin{
+	if viper.GetBool(constants.EnvFakeNode) {
+		return &FakeNodeDevicePlugin{
+			kubeClient: kubeClient,
+			gpuCount:   getGpuCount(topology),
+		}
+	}
+
+	return &RealNodeDevicePlugin{
 		devs:   createDevices(getGpuCount(topology)),
 		socket: serverSock,
 	}
-}
-
-func getGpuCount(nodeTopology *topology.NodeTopology) int {
-	return len(nodeTopology.Gpus)
-}
-
-func (m *DevicePlugin) GetDevicePluginOptions(context.Context, *pluginapi.Empty) (*pluginapi.DevicePluginOptions, error) {
-	return &pluginapi.DevicePluginOptions{}, nil
-}
-
-func dial(unixSocketPath string, timeout time.Duration) (*grpc.ClientConn, error) {
-	ctx := context.Background()
-	ctx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-
-	c, err := grpc.DialContext(
-		ctx,
-		unixSocketPath,
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithBlock(),
-		grpc.WithContextDialer(func(_ context.Context, addr string) (net.Conn, error) {
-			return net.DialTimeout("unix", addr, timeout)
-		}),
-	)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return c, nil
-}
-
-func createDevices(devCount int) []*pluginapi.Device {
-	var devs []*pluginapi.Device
-	for i := 0; i < devCount; i++ {
-		u, _ := uuid.NewRandom()
-		devs = append(devs, &pluginapi.Device{
-			ID:     u.String(),
-			Health: pluginapi.Healthy,
-		})
-	}
-	return devs
-}
-
-func (m *DevicePlugin) Start() error {
-	err := m.cleanup()
-	if err != nil {
-		return err
-	}
-
-	sock, err := net.Listen("unix", m.socket)
-	if err != nil {
-		return err
-	}
-
-	m.server = grpc.NewServer([]grpc.ServerOption{}...)
-	pluginapi.RegisterDevicePluginServer(m.server, m)
-
-	go func() {
-		err := m.server.Serve(sock)
-		if err != nil {
-			log.Println(err)
-		}
-	}()
-
-	// Wait for server to start by launching a blocking connexion
-	conn, err := dial(m.socket, 5*time.Second)
-	if err != nil {
-		return err
-	}
-	conn.Close()
-
-	return nil
-}
-
-func (m *DevicePlugin) Stop() error {
-	if m.server == nil {
-		return nil
-	}
-
-	m.server.Stop()
-	m.server = nil
-	close(m.stop)
-
-	return m.cleanup()
-}
-
-func (m *DevicePlugin) Register(kubeletEndpoint, resourceName string) error {
-	conn, err := dial(kubeletEndpoint, 5*time.Second)
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-
-	client := pluginapi.NewRegistrationClient(conn)
-	reqt := &pluginapi.RegisterRequest{
-		Version:      pluginapi.Version,
-		Endpoint:     path.Base(m.socket),
-		ResourceName: resourceName,
-	}
-
-	_, err = client.Register(context.Background(), reqt)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (m *DevicePlugin) ListAndWatch(e *pluginapi.Empty, s pluginapi.DevicePlugin_ListAndWatchServer) error {
-	err := s.Send(&pluginapi.ListAndWatchResponse{Devices: m.devs})
-	if err != nil {
-		fmt.Printf("Failed to send devices to Kubelet: %v\n", err)
-	}
-
-	for {
-		select {
-		case <-m.stop:
-			return nil
-		case d := <-m.health:
-			// FIXME: there is no way to recover from the Unhealthy state.
-			d.Health = pluginapi.Unhealthy
-			err := s.Send(&pluginapi.ListAndWatchResponse{Devices: m.devs})
-			if err != nil {
-				log.Printf("failed to send unhealthy update: %v", err)
-			}
-		}
-	}
-}
-
-func (m *DevicePlugin) GetPreferredAllocation(context.Context, *pluginapi.PreferredAllocationRequest) (*pluginapi.PreferredAllocationResponse, error) {
-	return &pluginapi.PreferredAllocationResponse{}, nil
-}
-
-func (m *DevicePlugin) Allocate(ctx context.Context, reqs *pluginapi.AllocateRequest) (*pluginapi.AllocateResponse, error) {
-	responses := pluginapi.AllocateResponse{}
-	for _, req := range reqs.ContainerRequests {
-		response := pluginapi.ContainerAllocateResponse{
-			Envs: map[string]string{
-				"MOCK_NVIDIA_VISIBLE_DEVICES": strings.Join(req.DevicesIDs, ","),
-			},
-			Mounts: []*pluginapi.Mount{
-				{
-					ContainerPath: "/bin/nvidia-smi",
-					HostPath:      "/var/lib/runai/bin/nvidia-smi",
-				},
-			},
-		}
-
-		responses.ContainerResponses = append(responses.ContainerResponses, &response)
-	}
-
-	return &responses, nil
-}
-
-func (m *DevicePlugin) PreStartContainer(context.Context, *pluginapi.PreStartContainerRequest) (*pluginapi.PreStartContainerResponse, error) {
-	return &pluginapi.PreStartContainerResponse{}, nil
-}
-
-func (m *DevicePlugin) cleanup() error {
-	if err := os.Remove(m.socket); err != nil && !os.IsNotExist(err) {
-		return err
-	}
-
-	return nil
-}
-
-func (m *DevicePlugin) Serve() error {
-	err := m.Start()
-	if err != nil {
-		log.Printf("Could not start device plugin: %s", err)
-		return err
-	}
-	log.Println("Starting to serve on", m.socket)
-
-	err = m.Register(pluginapi.KubeletSocket, resourceName)
-	if err != nil {
-		log.Printf("Could not register device plugin: %s", err)
-		stopErr := m.Stop()
-		if stopErr != nil {
-			log.Printf("Could not stop device plugin: %s", stopErr)
-		}
-		return err
-	}
-	log.Println("Registered device plugin with Kubelet")
-
-	return nil
 }
